Use network automation rule
